refactor(tui): modularize tui.rs into components and utils
Breaks down the large `tui.rs` monolith by introducing a `tui/` directory.
Rendering logic is now in `components.rs` and generic helpers in `utils.rs`,
leaving `mod.rs` to handle core App state and routing.
